Refs #124. Moved heading testing to its own module
