+++ tests/export/test_html.py
@@ -8,16 +8,15 @@ from __future__ import (
 
 import base64
 
+from pydocx.constants import EMUS_PER_PIXEL
+from pydocx.test import DocumentGeneratorTestCase
+from pydocx.test.utils import WordprocessingDocumentFactory
 from pydocx.wordml import (
     ImagePart,
     MainDocumentPart,
     NumberingDefinitionsPart,
     StyleDefinitionsPart,
 )
-
-from pydocx.constants import EMUS_PER_PIXEL
-from pydocx.test import DocumentGeneratorTestCase
-from pydocx.test.utils import WordprocessingDocumentFactory
 
 
 class FakedSubScriptTestCase(DocumentGeneratorTestCase):
@@ -318,99 +317,6 @@ class FakedSuperScriptTestCase(DocumentGeneratorTestCase):
         document.add(MainDocumentPart, document_xml)
 
         expected_html = '<p>nth</p>'
-        self.assert_document_generates_html(document, expected_html)
-
-
-class HeadingTestCase(DocumentGeneratorTestCase):
-    def test_character_stylings_are_ignored(self):
-        # Even though the heading1 style has bold enabled, it's being ignored
-        # because the style is for a header
-        style_xml = '''
-            <style styleId="heading1" type="paragraph">
-              <name val="Heading 1"/>
-              <rPr>
-                <b val="on"/>
-              </rPr>
-            </style>
-        '''
-
-        document_xml = '''
-            <p>
-              <pPr>
-                <pStyle val="heading1"/>
-              </pPr>
-              <r>
-                <t>aaa</t>
-              </r>
-            </p>
-        '''
-
-        document = WordprocessingDocumentFactory()
-        document.add(StyleDefinitionsPart, style_xml)
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '''
-            <h1>aaa</h1>
-        '''
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_each_heading_level(self):
-        style_template = '''
-            <style styleId="heading%s" type="paragraph">
-              <name val="Heading %s"/>
-            </style>
-        '''
-
-        style_xml = ''.join(
-            style_template % (i, i)
-            for i in range(1, 11)
-        )
-
-        paragraph_template = '''
-            <p>
-              <pPr>
-                <pStyle val="%s"/>
-              </pPr>
-              <r>
-                <t>%s</t>
-              </r>
-            </p>
-        '''
-
-        style_to_text = [
-            ('heading1', 'aaa'),
-            ('heading2', 'bbb'),
-            ('heading3', 'ccc'),
-            ('heading4', 'ddd'),
-            ('heading5', 'eee'),
-            ('heading6', 'fff'),
-            ('heading7', 'ggg'),
-            ('heading8', 'hhh'),
-            ('heading9', 'iii'),
-            ('heading10', 'jjj'),
-        ]
-
-        document_xml = ''.join(
-            paragraph_template % entry
-            for entry in style_to_text
-        )
-
-        document = WordprocessingDocumentFactory()
-        document.add(StyleDefinitionsPart, style_xml)
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '''
-            <h1>aaa</h1>
-            <h2>bbb</h2>
-            <h3>ccc</h3>
-            <h4>ddd</h4>
-            <h5>eee</h5>
-            <h6>fff</h6>
-            <h6>ggg</h6>
-            <h6>hhh</h6>
-            <h6>iii</h6>
-            <h6>jjj</h6>
-        '''
         self.assert_document_generates_html(document, expected_html)
